[0.6.0] Update ticket's power.

+++ Install/Database.sql
@@ -56,7 +56,7 @@ create table tardis_tickets
     title varchar(250) not null,
     start time not null,
     finish time not null,
-    power tinyint unsigned default 1 not null,
+    power smallint unsigned default 1 not null,
     constraint tardis_tickets_pk
         primary key (key_ticket),
     constraint tardis_tickets_tardis_lessons_key_lesson_fk
